requirements.txt
bokeh==2.4.0
dash-cytoscape==0.3.0
networkx==2.6.3
pandas==1.3.3
jupyter-dash==0.4.0